+++ index.ts
@@ -1,5 +1,6 @@
 export * from './lib/track';
 export * from './lib/api';
+export * from './lib/pipelines';
 export * from './lib/regions';
 export * from './lib/types';
 export { CustomerIORequestError } from './lib/utils';